Move `absolutize_relative_to` into `scolopasta-path` from `artichocke-backend`.
