+++ artichoke-backend/src/load_path/native.rs
@@ -5,9 +5,7 @@ use std::io;
 use std::path::{Path, PathBuf};
 
 use bstr::{BString, ByteSlice};
-use scolapasta_path::{normalize_slashes, ConvertBytesError};
-
-use super::absolutize_relative_to;
+use scolapasta_path::{absolutize_relative_to, normalize_slashes, ConvertBytesError};
 
 #[derive(Default, Debug, PartialEq, Eq)]
 pub struct Native {
