+++ scolapasta-path/src/lib.rs
@@ -34,7 +34,10 @@
 mod paths;
 mod platform_string;
 
-pub use paths::{is_explicit_relative, is_explicit_relative_bytes, memory_loader_ruby_load_path, normalize_slashes};
+pub use paths::{
+    absolutize_relative_to, is_explicit_relative, is_explicit_relative_bytes, memory_loader_ruby_load_path,
+    normalize_slashes,
+};
 pub use platform_string::{
     bytes_to_os_str, bytes_to_os_string, os_str_to_bytes, os_string_to_bytes, ConvertBytesError,
 };
